Moved utils folder and migrated fucntions from WolfracmCalc to utils/index.js
