--- React-Calculator/src/utils/index.js
@@ -1,8 +1,0 @@
-export const operations = {
-    '+': (x, y) => parseFloat(x) + parseFloat(y),
-    '-': (x, y) => parseFloat(x) - parseFloat(y),
-    '*': (x, y) => parseFloat(x) * parseFloat(y),
-    '/': (x, y) => parseFloat(x) / parseFloat(y),
-    '%': (x, y) => parseFloat(y) * (parseFloat(x) / 100),
-  };
-export const operators = Object.keys(operations);
